Moved SiteColorsPicker to homepage folder
